cmsis/hal drivers

+++ src/hardware/include/ltdc.h
@@ -32,7 +32,7 @@ typedef struct {
 	uint16_t y;
 	uint16_t width;
 	uint16_t height;
-    uint32_t transp;
+    uint8_t transp;
 	void * framebuf;
 } LTDC_Layer;
